Start re-thinking how configurations are stored
When this was running on a microcontroller I had an idea of having
a "running config" and a "saved config" that I would change via the USB
CDC device on the Pi Pico. This turned out to be more trouble that it's
worth.

Now that this code is running on Linux, and I have access to the file
system, each creature will be defined via a JSON file and read in when
the controller process starts.

This allows me to change the configuration on the fly without having to
do the build ➡️  flash ➡️  debug dance. (Or at least I hope it will!) It
also allows me to track changes over time.

The code built into the controller should work for any of my creatures,
and the only thing that changes is the config file.