Start to clean up the configuration / creature mess
My previous vision here was to create a split with the creature and the
configuration, kept in the Pi Pico's EEPROM. I wanted to be able to do
a Cisco IOS-like thing where I could edit the creature's config on the
fly in tihe console and the save it to the EEPROM on the Pico. This
didn't work out.

I'm trying to unwind this mess, and put it all in the `Creature` class.
The `Creature` class will be constructed from the `.json` file that's
passed in as a command line parameter when the controller starts. This
allows me to keep the controller code generic, and only change the
config file as needed.

That brings with it all the benefits of this sort of arrangement,
including being able to make changes using the Linux file system,
avoiding the build -> flash dance, and being able to track changes in
a git repo.