Phase 1 of abstracting out the Logger
I've hit a problem with the way I've been using the spdlog API directly
(it all comes from `namespace-stuffs.h`, which is a bad anti-pattern,
and I knew it was at the time) because it's very hard to mock out for
testing.

This is the first round of starting to abstract out the logger into an
interface that can be mocked in testing, and could be swapped out to
a different logger if I want to.

I also have an idea of making two loggers. One for the controller and
one for the firmware, but I haven't implemented that yet.