Don't move the logger, copy it
Newbie mistake. I don't want to `std::move()` the logger, I want to copy
it. If I move it the reference that passed in will go null (because
I told it to), which means that when it's used again in the parent class
it'll be null, and segfault.

Thank you, unit tests, for helping me find this. I'm sorry I doubted you
eariler! 💜