Start processing messages from the controller on the firmware
This is a big update that took a lot of hours to debug! We can now send
messages to the firmware from the controller and have it process them.
To do this we're using some C pointer crimes, but that's just how it has
to go to be done fast on an embedded system. 😅