Move the PWM counter logic back to the Pi Pico
This is important to me. The controller itself really only needs to know
about the number of microseconds the PWM pulse should be and nothing
more. It does not need to know (or care) about how the PWM subsystem
works.

Right now I have no plans to target anything other than the RP2040 for
the firmware, but that might always be the case. The RP2020 is a few
years old now, and no doubt RPI is already working on a new version. Or
maybe I'll want to use another type of MCU all together.

I've now verified with my scope that the PWM pulses are matching what's
in the config file, AND that they match what the old code was doing.
I now have no issues having the new controller control Beaky, for
example.

But I need to get some better error handling in place. It's all coming
together, however, and I'm super happy to be here!