Have the firmware request configuration, have controller send it
This is a HUGE update. I should have checked in my code more often, but
I was on a roll.

This change adds several new messages to the pipeline. The firmware will
refuse to process messages until it's gotten a configuration from the
controller, and the controller listens for and will respond to an INIT
message.

The firmware is now aware of when it's been connected to by
a controller. If the controller disconnects it will go back into an
unconfigured state, to help prevent mistakes when the controller changes
the configuration.

The firmware ensures that all movement commands are now within the min
and max spec, too. It will refuse to do anything out of those ranges.

I completely re-did how the status lights work after learning that
floating point on the Pi Pico is NOT horrible. RPI has floating point
instructions built into its ROM that are way faster than the ones gcc
supplies.

Now the DMX light is a status light that says what state the firmware is
in by the color. The next revision of the hardware will have that
corrected on the solder mask.