Allow control over UART in addition to USB
This is a really big refactoring of how messages are passed around.
This new config allows for messages to be passed over a UART and USB.
(Or both at the same time, but I don't recommend it!)

I'm thinking about making a Pi HAT and needed a way for the Pi to talk
to the Pico without using a tiny USB cable. This might do the trick. I'm
gonna find out.