Allow the controller to reset the firmware's buffers 🔔
This adds a new `FlushBuffer` message, but that message is special since
it's a single character. That character happens to be `\a`, which is
a bell. 🔔 This character is a magic character that tells the firmware
it's time to flush the incoming buffer and get ready for new work.

This was needed to allow the firmware to reset everything when the
controller restarts.

I'm getting a bit concerned about how much memory this new message
pipeline uses. I'm down to 35k free heap, which is fine (there's no
sense in not using all of the memory on the device), but that's cutting
it really close.

The extra memory is because the outgoing buffer needs to be fairly deep.
If it's not, messages get lost, like the message saying that the
firmware is ready. If memory gets too tight I will QoS and allow a READY
message to jump to the head of the queue, dropping other messages if
needed. That would let me back the size of the queues back down a lot.