Give up on the PCA9685
It's a good chip, and for most people it'll be just fine. But it's made
for LED lights and not animatronics. I can't get it to spit out
a consistent 50Hz.

Most of the time it's high, around 54Hz. I was able to mess around with
the prescaler to get it somewhat close to 50Hz, but that's not great. It
wasn't matching the datasheet.

I need consistent performance since this is for controlling servos and
servos are very sensative to this stuff. I'm nervous that it'll vary
from chip to chip.

I'm not going to delete this code, and might revisit it in the future,
but for now, the road ends here.