Switch from the `ads1115` to the `mcp3304`
The fastest I can get a reading off the `ads1115` is 2ms. That's far too
long for what I need, since I need to be able to poll up to eight motors
at once and fit it all within 10ms.

Soooooo I switched over the `mcp3304`. This is the same ADC I use on my
joysticks, and I know they work really well. SPI is a lot faster in this
use case, leaving the i2c bus for things that don't need to be read
nearly as fast, like the temperature.

This comes at a cost, however. I needed to steal four pins that I had
earmarked for stepper control. I've never actually use that
functionality, but I was still setting it aside for later. If I need to,
I'll re-think this later, but this works for now.