Allow the build to run on a RP2350 (Pico 2)
There's a few changes that I needed to make. One, I needed to allow
myself to completely disable the sensors, because the RP2350 boards
I have do not have sensors on them. (They're just devkits.)

Second, `configRUN_FREERTOS_SECURE_ONLY` needs to be set to 1 to work on
the RP2350. I'm not sure what that is, but it sounds good, so okay!